staticcheck
AO-17813
