+++ v1/plugin/mocks_test.go
@@ -34,7 +34,6 @@
 package plugin
 
 import (
-	"context"
 	"errors"
 	"fmt"
 	"time"
@@ -46,14 +45,6 @@ type mockPlugin struct {
 	err error
 }
 
-func newMockPlugin() *mockPlugin {
-	return &mockPlugin{}
-}
-
-func newMockErrPlugin() *mockPlugin {
-	return &mockPlugin{err: errors.New("error")}
-}
-
 type mockPublisher struct {
 	mockPlugin
 	err       error
@@ -62,10 +53,6 @@ type mockPublisher struct {
 
 func newMockPublisher() *mockPublisher {
 	return &mockPublisher{}
-}
-
-func newMockErrPublisher() *mockPublisher {
-	return &mockPublisher{err: errors.New("error")}
 }
 
 func (mpb *mockPublisher) Publish(mts []Metric, cfg Config) error {
@@ -99,55 +86,6 @@ func (mp *mockPlugin) GetConfigPolicy() (ConfigPolicy, error) {
 	return (*cp), nil
 }
 
-type mockStreamer struct {
-	mockPlugin
-	err                error
-	maxBuffer          int64
-	maxCollectDuration time.Duration
-	inMetric           chan []Metric
-	outMetric          chan []Metric
-	action             func(chan []Metric, time.Duration, []Metric)
-}
-
-func newMockStreamer() *mockStreamer {
-	return &mockStreamer{}
-}
-
-func newMockErrStreamer() *mockStreamer {
-	return &mockStreamer{err: errors.New("empty")}
-}
-
-func newMockStreamerStream(action func(chan []Metric, time.Duration, []Metric)) *mockStreamer {
-	return &mockStreamer{action: action}
-}
-
-func (mc *mockStreamer) doAction(t time.Duration, mts []Metric) {
-	go func() {
-		mc.action(mc.outMetric, t, mts)
-	}()
-}
-func (mc *mockStreamer) GetMetricTypes(cfg Config) ([]Metric, error) {
-	if mc.err != nil {
-		return nil, errors.New("error")
-	}
-
-	mts := []Metric{}
-	for _, v := range getMockMetricDataMap() {
-		mts = append(mts, v)
-	}
-	return mts, nil
-}
-
-func (mc *mockStreamer) StreamMetrics(ctx context.Context, i chan []Metric, o chan []Metric, _ chan string) error {
-
-	if mc.err != nil {
-		return errors.New("error")
-	}
-	mc.inMetric = i
-	mc.outMetric = o
-	return nil
-}
-
 type mockCollector struct {
 	mockPlugin
 	err              error
@@ -157,10 +95,6 @@ type mockCollector struct {
 
 func newMockCollector() *mockCollector {
 	return &mockCollector{}
-}
-
-func newMockErrCollector() *mockCollector {
-	return &mockCollector{err: errors.New("empty")}
 }
 
 func (mc *mockCollector) GetMetricTypes(cfg Config) ([]Metric, error) {
@@ -195,10 +129,6 @@ type mockProcessor struct {
 
 func newMockProcessor() *mockProcessor {
 	return &mockProcessor{}
-}
-
-func newMockErrProcessor() *mockProcessor {
-	return &mockProcessor{err: errors.New("error")}
 }
 
 func (mp *mockProcessor) Process(mts []Metric, cfg Config) ([]Metric, error) {
